\name{data_NGHS}
\alias{data_NGHS}
\docType{data}
\title{a longitudinal BMI and systolic BP data from the NGHS website}

\format{
  A data frame with 19701 observations on the following 11 variables.
}

\source{
  https: //biolincc.nhlbi.nih.gov/
}
\references{
  
}
\examples{
  
}
\keyword{datasets}
